pyproject.toml
[tool.poetry]
name = "iis-temperature-forecaster"
version = "0.1.0"
description = "Temperature forecaster intelligent system"
authors = ["JakobOpresnik <jakob.opresnik@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = ">=1.2.4,<2.0.0"
pandas = ">=2.2.3,<3.0.0"
scikit-learn = ">=1.6.1,<2.0.0"
requests = ">=2.32.3,<3.0.0"
lxml = ">=5.3.1,<6.0.0"
pyaml = ">=25.1.0,<26.0.0"
great-expectations = "==0.18.21"
evidently = ">=0.7.5,<0.8.0"
torch = ">=2.7.0,<3.0.0"
matplotlib = ">=3.10.3,<4.0.0"
mlflow = ">=2.22.0,<3.0.0"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.3,<0.35.0"
supabase = "^2.15.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"